🧑‍🔬 Allow programs to extend the EReq and ERes types used within the Router and support type inference in the Router methods
Previously typescript could not infer the types used in the callbacks on the Router methods `use`, `get` etc

This patch brings supports for typescript to infer types in the callbacks correctly and also brings support for applications to extend those types.

What this means is we can now implement applications which add new methods to the Request or Response, and have those new methods be understood by our type signatures.

E.G - this is now all type-checked correctly:
```ts
interface MeowReq extends EReq {
    meow(): string
  }

  interface WoofReq extends EReq {
    meow(): string
  }

  interface MyReq extends MeowReq, WoofReq { }

  interface WoofRes extends ERes {
    woof(): string
  }

  const router = new Router<MyReq, WoofRes>();
  router.all('/', async (req, res) => {
    req.meow()
    res.woof()
    return res.send("Hello world!");
  });
```